+++ packages/react/src/hooks/useChains.ts
@@ -6,14 +6,14 @@ import { ChainData, ChainType } from '../types/index.js';
  * A hook that returns an array of supported chains.
  * If no type is provided, it returns all chains.
  * @param type The type of chain to return
- * @returns An array of `ChainData<type>`
+ * @returns An array of `ChainData`
  */
 export const useChains = <T extends ChainType>(type?: T): ChainData[] => {
   const { chains } = useContext(TangledContext);
 
   return useMemo(() => {
     if (type) {
-      return chains[type] as ChainData<T>[];
+      return chains[type] as ChainData[];
     }
 
     return Object.values(chains).flat();