+++ packages/react/src/hooks/useWaitForTransaction.ts
@@ -1,16 +1,18 @@
 import { useQuery } from '@tanstack/react-query';
 import { TransactionParams, WatchTransactionOverrides, waitForTransaction } from '../actions/waitForTransaction.js';
-import { ChainId, TransactionReceipt } from '../types/index.js';
+import { ChainId, ChainType } from '../types/index.js';
+import { QueryParameter } from '../types/properties.js';
 import { useChain } from './useChain.js';
 import { useConnectionOrConfig } from './useConnectionOrConfig.js';
 
 export type UseWaitForTransactionParams = {
   /** Transaction Parameters {@link TransactionParams} */
-  transactionParams: TransactionParams;
+  transactionParams: TransactionParams<ChainType>;
   /** Chain ID of transaction */
   chainId: ChainId | undefined;
   /** Transaction overrides {@link WatchTransactionOverrides} */
-  overrides?: WatchTransactionOverrides;
+  overrides?: WatchTransactionOverrides<ChainType>;
+  queryOptions?: QueryParameter;
 };
 
 /**
@@ -18,11 +20,16 @@ export type UseWaitForTransactionParams = {
  * @param transactionParams - Transaction Parameters
  * @param chainId - Chain ID of transaction
  */
-export const useWaitForTransaction = ({ transactionParams, chainId, overrides }: UseWaitForTransactionParams) => {
+export const useWaitForTransaction = ({
+  transactionParams,
+  chainId,
+  overrides,
+  queryOptions = {},
+}: UseWaitForTransactionParams) => {
   const connectionOrConfig = useConnectionOrConfig();
   const chain = useChain(chainId);
 
-  return useQuery<TransactionReceipt>({
+  return useQuery({
     queryKey: ['watch transaction', transactionParams, chainId],
     queryFn: async () => {
       if (!connectionOrConfig) {
@@ -45,5 +52,6 @@ export const useWaitForTransaction = ({ transactionParams, chainId, overrides }:
     enabled: Boolean(connectionOrConfig && chain && transactionParams),
     retry: false,
     refetchOnWindowFocus: false,
+    ...queryOptions,
   });
 };