+++ packages/react/src/utils/getDefaultSupportedChains.ts
@@ -3,7 +3,7 @@ import * as evm from '../chains/evm.js';
 import { solana } from '../chains/solana.js';
 import { tonMainnet } from '../chains/ton.js';
 import { tronMainnet } from '../chains/tron.js';
-import { ChainData, SupportedChainsByType } from '../types/index.js';
+import { EVMChain, OtherChainData, SupportedChainsByType, TronChain } from '../types/index.js';
 
 const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => {
   const supportedChains: SupportedChainsByType = {
@@ -41,11 +41,11 @@ const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => 
       evm.polygonZkEvm,
       evm.scroll,
       evm.zkSync,
-    ] as ChainData<'evm'>[];
-    supportedChains.solana = [solana] as ChainData<'solana'>[];
-    supportedChains.tron = [tronMainnet] as ChainData<'tron'>[];
-    supportedChains.alephZero = [alephZero] as ChainData<'alephZero'>[];
-    supportedChains.ton = [tonMainnet] as ChainData<'ton'>[];
+    ] as EVMChain[];
+    supportedChains.solana = [solana] as OtherChainData<'solana'>[];
+    supportedChains.tron = [tronMainnet] as TronChain[];
+    supportedChains.alephZero = [alephZero] as OtherChainData<'alephZero'>[];
+    supportedChains.ton = [tonMainnet] as OtherChainData<'ton'>[];
   }
 
   return supportedChains;