+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tangled3/react
 
+## 1.6.0
+
+### Minor Changes
+
+- refactor chain types and add token handler hooks
+
+## 1.5.3
+
+### Patch Changes
+
+- fix send transaction mutation arguments
+
 ## 1.5.2
 
 ### Patch Changes